I set up my account at web1on1, created a new bot and got the neccessary API and webhook keys to be able to work.
I also set up a crude express server to verify the webhook and to hook a chipchat instance to it.
Comment: I wasn't able to suer the chipchat as middleware, but the "raw ingest" approach seems to be working for now.

I set up a very basic bot structure just to see if I am getting messages correctly.